+++ sample/src/main/java/app/cash/paparazzi/sample/util/SimpleTheme.kt
@@ -1,10 +1,9 @@
-package app.cash.paparazzi.sample
+package app.cash.paparazzi.sample.util
 
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
-
 
 enum class DesignTheme {
   LIGHT,
